Fixed Topbar's bug with logout; Changed Sidebar; Minor refactors of hooks and  pages; Added logo to public folder and icon in browser tab; Renamed Lounge page to Lessons
