+++ public/index.html
@@ -7,7 +7,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Roboto+Mono&display=swap" rel="stylesheet">
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <link rel="icon" href="%PUBLIC_URL%/MnC_logo_small.png" />
+    <link rel="icon" href="%PUBLIC_URL%/logo.svg" />
     <title id="htmlTitle">M&C</title>
   </head>
   <body>
