+++ src/components/Sidebar/Sidebar.module.css
@@ -12,7 +12,7 @@
 }
 
 .sidebarImage {
-    margin: 10px 0;
+    margin: 10px 0 10px;
     display: flex;
     justify-content: center;
 }
