+++ src/components/UI/MyModal/MyModal.module.css
@@ -7,7 +7,7 @@
     left: 0;
     right: 0;
     background: rgba(0, 0, 0, 0.7);
-    transition: opacity 0.5s ease;
+    transition: opacity 0.3s ease, visibility 0.3s ease;
     z-index: 99;
 }
 
@@ -23,7 +23,6 @@
     border-radius: 5px;
     width: 30%;
     position: relative;
-    transition: all 0.5s ease;
 }
 
 @media screen and (max-width: 700px){
