+++ src/hooks/useAuth.js
@@ -1,4 +1,4 @@
-import { useContext, useDebugValue } from "react";
+import {useContext, useDebugValue} from "react";
 import AuthContext from "../context/AuthProvider";
 
 const useAuth = () => {
