+++ src/hooks/useLogout.js
@@ -2,20 +2,20 @@ import axios from "../http/index";
 import useAuth from "./useAuth";
 
 const useLogout = () => {
+
     const { setAuth } = useAuth();
 
-    const logout = async () => {
-        setAuth({});
+    return async () => {
+        setAuth({})
         try {
             const response = await axios('/logout', {
                 withCredentials: true
             });
+
         } catch (err) {
             console.error(err);
         }
-    }
-
-    return logout;
+    };
 }
 
-export default useLogout
+export default useLogout;
