+++ src/pages/Lessons.jsx
@@ -1,0 +1,12 @@
+import Topbar from "../components/Topbar/Topbar";
+import Users from "../components/Users";
+
+const Lessons = () => {
+    return (
+        <Topbar title={'Уроки'}>
+            <Users />
+        </Topbar>
+    )
+}
+
+export default Lessons;
