+++ src/styles/App.css
@@ -28,8 +28,8 @@ p {
     margin: 5px 5px;
 }
 
-.login{
-    background: linear-gradient(to right top, #8D99AE 0%, #EDF2F4 100%);
+.formBackground{
+    background: linear-gradient(to left bottom, #8D99AE 0%, #EDF2F4 40%);
     width: 100%;
     height: 100vh;
     overflow-y: scroll;
@@ -110,12 +110,13 @@ a, a:hover, a:focus {
     flex-direction: column;
     justify-content: center;
     align-items: center;
-    height: 85vh;
+    height: 100vh;
 }
 
 .notFoundLogo {
-    margin-right: 30px;
-    scale: 80%;
+    scale: 2.5;
+    margin-bottom: 50px;
+    margin-right: 10px;
 }
 
 .unAuthorized {
@@ -128,8 +129,9 @@ a, a:hover, a:focus {
 }
 
 .unAuthorizedLogo {
-    margin-right: 30px;
-    scale: 80%;
+    scale: 2.5;
+    margin-bottom: 50px;
+    margin-right: 10px;
 }
 
 .ItemList {
